// src/config/app-config.type.ts
export type AppConfig = {
  port: number;
  apiPrefix: string;
  geoLocationApi: string;
  geoLocationApiKey: string;
  weatherApi: string;
  weatherApiKey: string;
  timezoneApi: string;
  timezoneApiKey: string;
  newsApi: string;
  newsApiKey: string;
  currencyApi: string;
  currencyApiKey: string;
};